Move Content class to separate file
